# internal/swag/swagger.yaml
basePath: /
definitions:
...
  handler.ErrorResponse:
    properties:
      detail:
        type: string
      error:
        type: string
    type: object
...
host: api.trackride.app
info:
  contact:
    name: TrackRide Support
    url: https://github.com/nashirabbash/trackride
  description: Backend API for the TrackRide mobile application. Provides user authentication,
    vehicle tracking, ride logging, and social features.
  license:
    name: MIT
    url: https://github.com/nashirabbash/trackride/blob/master/LICENSE.txt
  title: TrackRide API
  version: 1.0.0
paths:
...
  /v1/feed:
    get:
      description: Retrieve rides from users you are following, with pagination and
        user's like status
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page, max 100 (default: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user feed
      tags:
      - Social
...
  /v1/rides/{id}/comments:
    post:
      consumes:
      - application/json
      description: Add a comment to a ride. If the ride owner has push notifications
        enabled, a notification is sent.
      parameters:
      - description: Ride ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      - description: Comment data (content required, max 280 chars)
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Ride not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Comment on a ride
      tags:
      - Social
  /v1/rides/{id}/like:
    post:
      description: Like another user's ride. If this is a new like and the ride owner
        has push notifications enabled, a notification is sent.
      parameters:
      - description: Ride ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Ride not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Like a ride
      tags:
      - Social
...
  /v1/users/{id}/follow:
    delete:
      description: Stop following another user
      parameters:
      - description: User ID to unfollow (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unfollow a user
      tags:
      - Social
    post:
      description: Follow another user to see their rides in your feed
      parameters:
      - description: User ID to follow (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Invalid ID or cannot follow self
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Follow a user
      tags:
      - Social
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token authentication. Pass as "Authorization: Bearer
      <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"